feature/weather-domain - console command, dependency injection, openweatherapi search implementation
