+++ app/composer.json
@@ -2,6 +2,10 @@
   "require": {
     "guzzlehttp/guzzle": "^7.0",
     "symfony/dotenv": "^5.3",
-    "symfony/console": "^5.3"
+    "symfony/console": "^5.3",
+    "symfony/dependency-injection": "^5.3",
+    "symfony/finder": "^5.3",
+    "symfony/config": "^5.3",
+    "symfony/yaml": "^5.3"
   }
 }
