+++ app/config/symfony/services.yaml
@@ -1,0 +1,4 @@
+services:
+  _defaults:
+    autowire: true
+    autoconfigure: true
