+++ app/src/Shared/Application/Query.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+namespace Shared\Application;
+
 interface Query
 {
