+++ app/src/Weather/Application/WeatherForCityQuery.php
@@ -2,6 +2,10 @@
 
 declare(strict_types=1);
 
+namespace Weather\Application;
+
+use Shared\Application\Query;
+
 final class WeatherForCityQuery implements Query
 {
     public function __construct(private string $cityName){}
