+++ app/src/Weather/Domain/WeatherRepository.php
@@ -2,7 +2,11 @@
 
 declare(strict_types=1);
 
+namespace Weather\Domain;
+
+use Weather\Application\WeatherForCityQuery;
+
 interface WeatherRepository
 {
-    public function searchForCity(WeatherForCityQuery $query);
+    public function searchForCity(WeatherForCityQuery $query): WeatherForCity;
 }
