+++ app/src/Weather/Infrastructure/ApiWeatherRepository.php
@@ -2,6 +2,13 @@
 
 declare(strict_types=1);
 
+namespace Weather\Infrastructure;
+
+use Shared\Infrastructure\ApiClient;
+use Weather\Application\WeatherForCityQuery;
+use Weather\Domain\WeatherForCity;
+use Weather\Domain\WeatherRepository;
+
 abstract class ApiWeatherRepository implements WeatherRepository
 {
     public function __construct(private ApiClient $client) {}
