+++ app/src/Weather/Infrastructure/OpenWeatherApiRepository.php
@@ -2,6 +2,12 @@
 
 declare(strict_types=1);
 
+namespace Weather\Infrastructure;
+
+use Shared\Infrastructure\ApiClient;
+use Weather\Application\WeatherForCityQuery;
+use Weather\Domain\WeatherForCity;
+
 final class OpenWeatherApiRepository extends ApiWeatherRepository
 {
     public function __construct(private ApiClient $client, private string $apiUrl, private string $apiKey)
